requirements.txt
requests>=2.32.3
python-dotenv>=1.0.0
crawl4ai>=0.5.0.post8
redis>=5.2.1
fastapi>=0.115.9
uvicorn>=0.34.0
pydantic>=2.11.2
starlette>=0.36.3
python-multipart>=0.0.6
aiohttp>=3.8.4
asyncio>=3.4.3
prometheus-client>=0.16.0
beautifulsoup4>=4.12.0
lxml>=4.9.2
schedule>=1.2.2
pytest>=7.3.1
httpx>=0.26.0
boto3>=1.37.28
playwright>=1.51.0
azure-storage-blob>=12.16.0
sentence-transformers>=2.2.2
chromadb~=0.4.24
pydantic-settings>=2.0.0
torch>=2.3.0,<2.7.0
torchvision>=0.18.0
# torchaudio>=2.2.0 # Removed unless specifically needed and compatible
langchain>=0.2.7,<0.3.0
langsmith>=0.1.75,<0.2.0
langchain-core>=0.2.10
langchain-chroma>=0.1.2
langchain-community>=0.2.7
# autoawq==0.2.6 # REMOVED
# autoawq-kernels==0.0.7 # REMOVED
importlib-metadata>=4.4
transformers>=4.30.0
tenacity==9.1.2